pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "knetvis"
version = "0.1.0"
authors = [
    { name = "Samuel Arogbonlo", email = "sbayo971@gmail.com" },
]
description = "A CLI tool for Kubernetes Network Policy visualization"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "kubernetes>=28.1.0",
    "networkx>=3.1",
    "matplotlib>=3.7.1",
    "click>=8.1.3",
    "rich>=13.3.5",
    "pyyaml>=6.0.1",
]

[project.scripts]
knetvis = "knetvis.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]